modify z to float in model forward

+++ model.py
@@ -49,7 +49,7 @@ class HOA_WavTokenizer(nn.Module):
         z = self.encoder(x)
 
         with torch.autocast(device_type=x.device.type, enabled=False):
-            vq_result = self.vq(z, frame_rate=75, bandwidth=bandwidth)
+            vq_result = self.vq(z.float(), frame_rate=75, bandwidth=bandwidth)
 
         decoded = self.decoder(vq_result.quantized)
         audio = self.head(decoded)
